add support for datefmt in FakeLogger
The logging fixture is extremely useful to be used a temp buffer for
collecting log messages into a buffer, and only decide if we're going
to emit them after some event in the future (like the failure or
success of some future criteria). However, in it's current form we are
not given access to the datefmt variable of the underlying Formatter,
which means we always end up with the default python time string for
%(asctime), which looks incorrectly localized many places.

This merely adds the ability to pass the datefmt param through to the
Formatter.

Signed-off-by: Sean Dague <sean@dague.net>