Python 3 MonkeyPatch staticmethod
In Python setattr(Class, name, func) automatically converts a function
into an instancemethod. To keep type(Class.func) as function,
staticmethod(func) must be applied explicitly. This was previously fixed
for Python 2 but the issue exists for 3 as well.

The test for this case was updated to correctly check it. The method
does not become bound until the class it is defined on is instantiated.