Loosen the requirements on MonkeyPatch methods
A previous change added some logic so that when monkeypatching a
staticmethod the old_value was restored as a staticmethod. As an
unintended side-effect this added a constraint that the method to patch
in had to match the patched method very closely. For example to patch a
method foo defined like:

class C(object):
    def foo(self):
        pass

you would need to patch foo with another instancemethod defined on class
C. The requirement should not be that strict.

This reworks the logic so that the following occurs:

* When patching C.foo with new_value check C.foo to see if it's a
  staticmethod and set new_value accordingly. This is a change as right
  now new_value is checked for being a staticmethod.
* When checking if a method is a staticmethod do it by checking C().foo
  because the bound/unbound distinction is not clear until C is
  instantiated.
* When restoring the original C.foo check that method to see if it's a
  staticmethod and restore it accordingly.

As a result staticmethods continue to be handled properly but it is now
possible to monkeypatch C.foo with a method baz(arg) defined anywhere.