Loosen the requirements on MonkeyPatch methods
A previous change added some logic so that when monkeypatching a
staticmethod the old_value was restored as a staticmethod. There was an
issue where the determination of whether or not a method should be
static was incorrectly checking the new function not the one to be
replaced. This caused an unintended side-effect that in order to patch
an instance method of a class the new function needed to be an instance
method of a class.

This reworks the logic so that the following occurs:

Given a class C defined as:

class C(object):
    def foo(self):
        pass

* When patching C.foo with new_value check C.foo to see if it's a
  staticmethod/classmethod and set new_value accordingly. This is a
  change as right now new_value is checked for being a staticmethod and
  classmethod was not handled at all.

* When calling addCleanup ensure that the original C.foo to be reset to
  remains a staticmethod/classmethod if it was originally. This is a
  change as classmethod was not handled.

As a result staticmethods continue to be handled properly but it is now
possible to monkeypatch C.foo with a method baz(arg) defined anywhere.