Add support for attaching logs on error.
Sometimes, code being tested can produce a large amount of output which
is needed to debug issues. Adding that output to successful tests can
lead to blowing out size maximums for a full test suite run, even if
only one test fails, because of adding the log streaming output to every
test all the time.

To empower users to only collect that output on failures, expose a flag
on the constructor which causes addOnException to be used instead of the
normal addDetail. The default stays at off so existing users should not
see a behavior change.