Fix MockPatchMultiple.DEFAULT
The change in #52 broke
`TestMockMultiple.test_mock_patch_without_replacement`, since
`MockPatchMultiple.DEFAULT` no longer worked as a class property.  Fix
this using a metaclass.

For compatibility, I arranged for it to also work as an instance
property; I suspect this isn't much needed since
`MockPatchMultiple.DEFAULT` is normally used as an argument to the
`MockPatchMultiple` constructor so you wouldn't normally have an
instance in hand, but it's easy enough to provide strict compatibility
here.