Move testtools requirement to a "streams" extra
Now that the core of `fixtures` no longer uses `testtools`, all that
remains to make it optional and break the cyclic requirement is to deal
with its use by streams fixtures.  These intrinsically rely on
`testtools`, but are also very much optional, so move the requirement to
a new `fixtures[streams]` extra.

Fixes #38.